+++ Examples/utilities/generate-pdf.robot
@@ -5,7 +5,10 @@ Test Setup    Open browser to test page
 Test Teardown    Close All Browser
 Suite Teardown    Close Puppeteer
 
+
 *** Variables ***
+${DEFAULT_BROWSER}    chrome
+# ${DEFAULT_BROWSER}    webkit
 ${HOME_PAGE_URL}    http://127.0.0.1:7272/basic-html-elements.html
 
 
@@ -14,7 +17,11 @@ Generate pdf file
     [Documentation]    Only support on headless mode
     Print as pdf 
 
+
 *** Keywords ***
 Open browser to test page
-    &{options} =    create dictionary   headless=${True}
-    Open browser    ${HOME_PAGE_URL}   options=${options}
+    ${BROWSER} =     Get variable value    ${BROWSER}    ${DEFAULT_BROWSER}
+    ${HEADLESS} =    Get variable value    ${HEADLESS}    ${False}
+    &{options} =    create dictionary   headless=${HEADLESS}
+    Open browser    ${HOME_PAGE_URL}    browser=${BROWSER}    options=${options}
+
