+++ PuppeteerLibrary/playwright/async_keywords/playwright_dropdown.py
@@ -1,4 +1,4 @@
-from PuppeteerLibrary.keywords.idropdown_async import iDropdownAsync
+from PuppeteerLibrary.ikeywords.idropdown_async import iDropdownAsync
 from PuppeteerLibrary.locators.SelectorAbstraction import SelectorAbstraction
