+++ PuppeteerLibrary/playwright/async_keywords/playwright_javascript.py
@@ -1,4 +1,4 @@
-from PuppeteerLibrary.keywords.ijavascript_async import iJavascriptAsync
+from PuppeteerLibrary.ikeywords.ijavascript_async import iJavascriptAsync
 
 
 class PlaywrightJavascript(iJavascriptAsync):
